requirements.txt
ipython
nose
path.py
PyYAML
